Feature/uz matrix without allocation issue 216

Approved-by: Dennis Hufnagel

+++ vitis/software/Baremetal/src/uz/uz_nn/uz_nn_layer.h
@@ -2,6 +2,9 @@
 #define UZ_NN_LAYER_H
 #include "../uz_matrix/uz_matrix.h"
 #include "uz_nn_activation_functions.h"
+#include <stdint.h>
+#include <stdint.h>
+
 
 /**
  * @brief Object definition for a layer of a neural network
@@ -25,11 +28,11 @@ enum activation_function{
  */
 struct uz_nn_layer_config{
     enum activation_function activation_function; /**< Activation function of all neurons in this layer */
-    size_t number_of_neurons; /**< Number of neurons in the layer */
-    size_t number_of_inputs; /**< Number of inputs to the layer. Is either the number of inputs to the network or the number of neurons of the previouse layer */
-    size_t length_of_weights; /**< Number of weights in the layer, has to be calculated by UZ_MATRIX_SIZE(weights) */
-    size_t length_of_bias; /**< Number of bias in the layer, has to be calculated by UZ_MATRIX_SIZE(bias) */
-    size_t length_of_output; /**< Number of outputs in the layer, has to be calculated by UZ_MATRIX_SIZE(output) and is equal to the number of weights */
+    uint32_t number_of_neurons; /**< Number of neurons in the layer */
+    uint32_t number_of_inputs; /**< Number of inputs to the layer. Is either the number of inputs to the network or the number of neurons of the previouse layer */
+    uint32_t length_of_weights; /**< Number of weights in the layer, has to be calculated by UZ_MATRIX_SIZE(weights) */
+    uint32_t length_of_bias; /**< Number of bias in the layer, has to be calculated by UZ_MATRIX_SIZE(bias) */
+    uint32_t length_of_output; /**< Number of outputs in the layer, has to be calculated by UZ_MATRIX_SIZE(output) and is equal to the number of weights */
     float *const weights; /** Pointer to an array that holds the weights */
     float *const bias; /** Pointer to an array that holds the bias */
     float *const output; /** Pointer to an array that holds the output / where the output is written to */
